+++ jenjin-world-client/src/main/java/com/jenjinstudios/world/message/ExecutableStateChangeMessage.java
@@ -54,7 +54,7 @@ public class ExecutableStateChangeMessage extends WorldClientExecutableMessage
 		Vector2D oldVector = new Vector2D(x,y);
 		double angle = MathUtil.calcStepAngle(absoluteAngle, relativeAngle);
 		double dist = ClientActor.MOVE_SPEED *
-				((System.nanoTime() - time) / 1000000000);
+				((double)(System.nanoTime() - time) / 1000000000d);
 		position = oldVector.getVectorInDirection(dist, angle);
 	}
 }
